+++ packages/ente-types/src/dtos/instance-config.dto.ts
@@ -1,0 +1,6 @@
+import { Languages } from "../languages";
+
+export class InstanceConfigDto {
+  defaultLanguage: Languages;
+  loginBanners: Record<Languages, string | null>;
+}